requirements.txt
alembic~=1.14.1
fastapi~=0.115.8
uvicorn~=0.34.0
gunicorn
sqlalchemy~=2.0.37
asyncpg
psycopg2-binary
python-dotenv~=1.0.1
fastapi-users[sqlalchemy]
fastapi-cache2[redis]
redis~=5.2.1
celery~=5.4.0